Carry out a resize even if no layout has been performed.
Add a FrameView resize in WebviewImpl::performResize if no layout has been
performed. This to fix the case where a layout is not performed because the
layout height is forced to zero (and therefore no layouts are performed when
the actual height is changed).

There were conflicts here -- some of the code that was supposed to be
edited in this commit in WebViewImpl::layoutUpdated had not been moved from
WebViewImpl::refreshPageScaleFactorAfterLayout to
WebViewImpl::layoutUpdated on this branch.

BUG=457159

Review URL: https://codereview.chromium.org/911083002

Change-Id: I2775bc0ba2684f914a9966a69cd580dbccd1e1a4
git-svn-id: svn://svn.chromium.org/blink/trunk@190424 bbb929c8-8fbe-4397-9dbb-9b2b20218538
Bug: 19105464